map collision optimisation

+++ Constants.h
@@ -2,8 +2,8 @@
 
 namespace Constants 
 {
-    const float pi{ 3.14159265359 };
-    const float g{ 30.0 };
+    const double pi{ 3.14159265359 };
+    const double g{ 30.0 };
 
     const int screenWidth{ 1600 };
     const int screenHeight{ 900 };
