+++ Player.h
@@ -9,22 +9,22 @@
 class Player : public Character
 {
 public:
-    Player(const char* fileName, float xStartPos = 0, float yStartPos = 0, float xVel = 0, float yVel = 0);
+    Player(const char* fileName, double xStartPos = 0, double yStartPos = 0, double xVel = 0, double yVel = 0);
     ~Player();
 
-    void cycleWalkAnimation(float timeStep);
-    void cycleIdleAnimation(float timeStep);
-    void spriteAnimate(float timeStep);
+    void cycleWalkAnimation(double timeStep);
+    void cycleIdleAnimation(double timeStep);
+    void spriteAnimate(double timeStep);
 
-    virtual void update(float timeStep, std::vector<std::vector<Tile>>& map) override;
+    virtual void update(double timeStep, std::vector<std::vector<Tile>>& map) override;
     virtual void draw() override;
 
     static const int spriteSheetCount{ 28 };
-    static constexpr float timeBetweenSpriteFrames{ 0.033333 };
+    static constexpr double timeBetweenSpriteFrames{ 0.033333 };
 
 private:
     std::array<SDL_Rect, spriteSheetCount> m_spriteRects;
     std::array<SDL_Rect, spriteSheetCount>::size_type m_spriteIndex;
     bool m_isFlipped{ false };
-    float m_animationTime{ 0 };
+    double m_animationTime{ 0 };
 };
